broke out editor for readability
