+++ package.json
@@ -33,6 +33,7 @@
   },
   "devDependencies": {
     "eslint": "7.32.0",
-    "eslint-config-next": "11.1.0"
+    "eslint-config-next": "11.1.0",
+    "prismjs": "^1.25.0"
   }
 }
