+++ src/components/Editor/Editor.js
@@ -1,32 +1,21 @@
 import React, { useCallback, useMemo } from "react";
-import { Slate, Editable, withReact, useSlate } from "slate-react";
-import {
-  Heading as ChakraHeading,
-  useColorMode,
-  List as ChakraList,
-  ListItem as ChakraListitem,
-  chakra,
-} from "@chakra-ui/react";
-import {
-  Editor,
-  isBoldActive,
-  isItalicActive,
-  Transforms,
-  Range,
-  Point,
-  createEditor,
-  Element as SlateElement,
-  Descendant,
-} from "slate";
+import Prism from "prismjs";
+
+import { Slate, Editable, withReact } from "slate-react";
+import Element from "./components/elements";
+import Leaf from "./components/leafs";
+import { BlockButton, MarkButton } from "./components/toggleButtons";
+import withShortcuts from "./editor-utils/withShortcut";
+
+import { Text, createEditor } from "slate";
 import { withHistory } from "slate-history";
 
-import { Toolbar, Icon } from "./EditorComponents/components";
+import { Toolbar } from "./components/components";
 
 import {
   FaBold,
   FaItalic,
   FaUnderline,
-  FaStrikethrough,
   FaCode,
   FaListOl,
   FaListUl,
@@ -34,20 +23,8 @@ import {
   FaHeading,
 } from "react-icons/fa";
 
-const LIST_TYPES = ["numbered-list", "bulleted-list"];
-
-const SHORTCUTS = {
-  "*": "list-item",
-  "-": "list-item",
-  "+": "list-item",
-  ">": "block-quote",
-  "#": "heading-one",
-  "##": "heading-two",
-  "###": "heading-three",
-  "####": "heading-four",
-  "#####": "heading-five",
-  "######": "heading-six",
-};
+// eslint-disable-next-line
+Prism.languages.markdown=Prism.languages.extend("markup",{}),Prism.languages.insertBefore("markdown","prolog",{blockquote:{pattern:/^>(?:[\t ]*>)*/m,alias:"punctuation"},code:[{pattern:/^(?: {4}|\t).+/m,alias:"keyword"},{pattern:/``.+?``|`[^`\n]+`/,alias:"keyword"}],title:[{pattern:/\w+.*(?:\r?\n|\r)(?:==+|--+)/,alias:"important",inside:{punctuation:/==+$|--+$/}},{pattern:/(^\s*)#+.+/m,lookbehind:!0,alias:"important",inside:{punctuation:/^#+|#+$/}}],hr:{pattern:/(^\s*)([*-])([\t ]*\2){2,}(?=\s*$)/m,lookbehind:!0,alias:"punctuation"},list:{pattern:/(^\s*)(?:[*+-]|\d+\.)(?=[\t ].)/m,lookbehind:!0,alias:"punctuation"},"url-reference":{pattern:/!?\[[^\]]+\]:[\t ]+(?:\S+|<(?:\\.|[^>\\])+>)(?:[\t ]+(?:"(?:\\.|[^"\\])*"|'(?:\\.|[^'\\])*'|\((?:\\.|[^)\\])*\)))?/,inside:{variable:{pattern:/^(!?\[)[^\]]+/,lookbehind:!0},string:/(?:"(?:\\.|[^"\\])*"|'(?:\\.|[^'\\])*'|\((?:\\.|[^)\\])*\))$/,punctuation:/^[\[\]!:]|[<>]/},alias:"url"},bold:{pattern:/(^|[^\\])(\*\*|__)(?:(?:\r?\n|\r)(?!\r?\n|\r)|.)+?\2/,lookbehind:!0,inside:{punctuation:/^\*\*|^__|\*\*$|__$/}},italic:{pattern:/(^|[^\\])([*_])(?:(?:\r?\n|\r)(?!\r?\n|\r)|.)+?\2/,lookbehind:!0,inside:{punctuation:/^[*_]|[*_]$/}},url:{pattern:/!?\[[^\]]+\](?:\([^\s)]+(?:[\t ]+"(?:\\.|[^"\\])*")?\)| ?\[[^\]\n]*\])/,inside:{variable:{pattern:/(!?\[)[^\]]+(?=\]$)/,lookbehind:!0},string:{pattern:/"(?:\\.|[^"\\])*"(?=\)$)/}}}}),Prism.languages.markdown.bold.inside.url=Prism.util.clone(Prism.languages.markdown.url),Prism.languages.markdown.italic.inside.url=Prism.util.clone(Prism.languages.markdown.url),Prism.languages.markdown.bold.inside.italic=Prism.util.clone(Prism.languages.markdown.italic),Prism.languages.markdown.italic.inside.bold=Prism.util.clone(Prism.languages.markdown.bold); // prettier-ignore
 
 export default function SlateEditor({ value, setValue }) {
   const renderElement = useCallback((props) => <Element {...props} />, []);
@@ -56,299 +33,98 @@ export default function SlateEditor({ value, setValue }) {
     () => withShortcuts(withReact(withHistory(createEditor()))),
     []
   );
-  return (
-    <Slate editor={editor} value={value} onChange={(value) => setValue(value)}>
-      <Toolbar>
-        <MarkButton format="bold" icon={<FaBold />} />
-        <MarkButton format="italic" icon={<FaItalic />} />
-        <MarkButton format="underline" icon={<FaUnderline />} />
-        <MarkButton format="code" icon={<FaCode />} />
-        <BlockButton format="heading-one" icon={<FaHeading />} />
-        <BlockButton format="block-quote" icon={<FaQuoteLeft />} />
-        <BlockButton format="numbered-list" icon={<FaListOl />} />
-        <BlockButton format="bulleted-list" icon={<FaListUl />} />
-      </Toolbar>
 
-      <Editable
-        renderElement={renderElement}
-        renderLeaf={renderLeaf}
-        placeholder="Write some markdown..."
-        spellCheck
-        autoFocus
-      />
-    </Slate>
+  const decorate = useCallback(([node, path]) => {
+    const ranges = [];
+
+    if (!Text.isText(node)) {
+      return ranges;
+    }
+
+    const getLength = (token) => {
+      if (typeof token === "string") {
+        return token.length;
+      } else if (typeof token.content === "string") {
+        return token.content.length;
+      } else {
+        return token.content.reduce((l, t) => l + getLength(t), 0);
+      }
+    };
+
+    const tokens = Prism.tokenize(node.text, Prism.languages.markdown);
+    let start = 0;
+
+    for (const token of tokens) {
+      const length = getLength(token);
+      const end = start + length;
+
+      if (typeof token !== "string") {
+        ranges.push({
+          [token.type]: true,
+          anchor: { path, offset: start },
+          focus: { path, offset: end },
+        });
+      }
+
+      start = end;
+    }
+
+    return ranges;
+  }, []);
+  return (
+    <>
+      <Slate
+        editor={editor}
+        value={value}
+        onChange={(value) => setValue(value)}
+      >
+        <Toolbar>
+          <MarkButton format="bold" icon={<FaBold />} />
+          <MarkButton format="italic" icon={<FaItalic />} />
+          <MarkButton format="underline" icon={<FaUnderline />} />
+          <MarkButton format="code" icon={<FaCode />} />
+          <BlockButton format="heading-one" icon={<FaHeading />} />
+          <BlockButton format="block-quote" icon={<FaQuoteLeft />} />
+          <BlockButton format="numbered-list" icon={<FaListOl />} />
+          <BlockButton format="bulleted-list" icon={<FaListUl />} />
+        </Toolbar>
+        <Editable
+          renderElement={renderElement}
+          renderLeaf={renderLeaf}
+          decorate={decorate}
+          placeholder="Write some markdown..."
+          autoFocus
+          onSelect={() => {
+            /**
+             * Chrome doesn't scroll at bottom of the page. This fixes that.
+             */
+            if (!window.chrome) return;
+            if (editor.selection == null) return;
+            try {
+              /**
+               * Need a try/catch because sometimes you get an error like:
+               *
+               * Error: Cannot resolve a DOM node from Slate node: {"type":"p","children":[{"text":"","by":-1,"at":-1}]}
+               */
+              const domPoint = ReactEditor.toDOMPoint(
+                editor,
+                editor.selection.focus
+              );
+              const node = domPoint[0];
+              if (node == null) return;
+              const element = node.parentElement;
+              if (element == null) return;
+              element.scrollIntoView({ behavior: "smooth", block: "nearest" });
+            } catch (e) {
+              /**
+               * Empty catch. Do nothing if there is an error.
+               */
+            }
+          }}
+        />
+      </Slate>
+
+      <h1> value: {JSON.stringify(value)} </h1>
+    </>
   );
 }
-
-const withShortcuts = (editor) => {
-  const { deleteBackward, insertText } = editor;
-
-  editor.insertText = (text) => {
-    const { selection } = editor;
-
-    if (text === " " && selection && Range.isCollapsed(selection)) {
-      const { anchor } = selection;
-      const block = Editor.above(editor, {
-        match: (n) => Editor.isBlock(editor, n),
-      });
-      const path = block ? block[1] : [];
-      const start = Editor.start(editor, path);
-      const range = { anchor, focus: start };
-      const beforeText = Editor.string(editor, range);
-      const type = SHORTCUTS[beforeText];
-
-      if (type) {
-        Transforms.select(editor, range);
-        Transforms.delete(editor);
-        const newProperties = {
-          type,
-        };
-        Transforms.setNodes(editor, newProperties, {
-          match: (n) => Editor.isBlock(editor, n),
-        });
-
-        if (type === "list-item") {
-          const list = {
-            type: "bulleted-list",
-            children: [],
-          };
-          Transforms.wrapNodes(editor, list, {
-            match: (n) =>
-              !Editor.isEditor(n) &&
-              SlateElement.isElement(n) &&
-              n.type === "list-item",
-          });
-        }
-
-        return;
-      }
-    }
-
-    insertText(text);
-  };
-
-  editor.deleteBackward = (...args) => {
-    const { selection } = editor;
-
-    if (selection && Range.isCollapsed(selection)) {
-      const match = Editor.above(editor, {
-        match: (n) => Editor.isBlock(editor, n),
-      });
-
-      if (match) {
-        const [block, path] = match;
-        const start = Editor.start(editor, path);
-
-        if (
-          !Editor.isEditor(block) &&
-          SlateElement.isElement(block) &&
-          block.type !== "paragraph" &&
-          Point.equals(selection.anchor, start)
-        ) {
-          const newProperties = {
-            type: "paragraph",
-          };
-          Transforms.setNodes(editor, newProperties);
-
-          if (block.type === "list-item") {
-            Transforms.unwrapNodes(editor, {
-              match: (n) =>
-                !Editor.isEditor(n) &&
-                SlateElement.isElement(n) &&
-                n.type === "bulleted-list",
-              split: true,
-            });
-          }
-
-          return;
-        }
-      }
-
-      deleteBackward(...args);
-    }
-  };
-
-  return editor;
-};
-
-const toggleBlock = (editor, format) => {
-  const isActive = isBlockActive(editor, format);
-  const isList = LIST_TYPES.includes(format);
-
-  Transforms.unwrapNodes(editor, {
-    match: (n) =>
-      LIST_TYPES.includes(
-        !Editor.isEditor(n) && SlateElement.isElement(n) && n.type
-      ),
-    split: true,
-  });
-  const newProperties = {
-    type: isActive ? "paragraph" : isList ? "list-item" : format,
-  };
-  Transforms.setNodes(editor, newProperties);
-
-  if (!isActive && isList) {
-    const block = { type: format, children: [] };
-    Transforms.wrapNodes(editor, block);
-  }
-};
-
-const toggleMark = (editor, format) => {
-  const isActive = isMarkActive(editor, format);
-
-  if (isActive) {
-    Editor.removeMark(editor, format);
-  } else {
-    Editor.addMark(editor, format, true);
-  }
-};
-
-const isBlockActive = (editor, format) => {
-  const [match] = Editor.nodes(editor, {
-    match: (n) =>
-      !Editor.isEditor(n) && SlateElement.isElement(n) && n.type === format,
-  });
-
-  return !!match;
-};
-
-const isMarkActive = (editor, format) => {
-  const marks = Editor.marks(editor);
-  return marks ? marks[format] === true : false;
-};
-
-const Element = ({ attributes, children, element }) => {
-  const { colorMode } = useColorMode();
-  const color = {
-    light: "primary",
-    dark: "highlight",
-  };
-
-  switch (element.type) {
-    case "block-quote":
-      return (
-        <chakra.blockquote
-          borderLeft="2px"
-          marginLeft="0"
-          marginRight="0"
-          paddingLeft="10px"
-          color="#aaa"
-          {...attributes}
-        >
-          {children}
-        </chakra.blockquote>
-      );
-    case "bulleted-list":
-      return (
-        <ChakraList
-          display="block"
-          marginBlockStart="1em"
-          marginBlockEnd="1em"
-          marginInlineStart="0px"
-          paddingInlineStart="40px"
-          listStyleType="disc"
-          {...attributes}
-        >
-          {children}
-        </ChakraList>
-      );
-    case "list-item":
-      return (
-        <ChakraListitem listStyleType="disc" fontSize="30px" {...attributes}>
-          {children}
-        </ChakraListitem>
-      );
-    case "heading-one":
-      return (
-        <ChakraHeading
-          as="h1"
-          color={color[colorMode]}
-          size="4xl"
-          lineHeight="2"
-          fontWeight="bold"
-          {...attributes}
-        >
-          {children}
-        </ChakraHeading>
-      );
-    case "heading-two":
-      return (
-        <ChakraHeading
-          as="h2"
-          color={color[colorMode]}
-          size="3xl"
-          lineHeight="2"
-          fontWeight="bold"
-          {...attributes}
-        >
-          {children}
-        </ChakraHeading>
-      );
-    case "heading-three":
-      return (
-        <ChakraHeading
-          as="h3"
-          color={color[colorMode]}
-          size="xl"
-          lineHeight="2"
-          fontWeight="bold"
-          {...attributes}
-        >
-          {children}
-        </ChakraHeading>
-      );
-
-    default:
-      return <p {...attributes}>{children}</p>;
-  }
-};
-
-const Leaf = ({ attributes, children, leaf }) => {
-  if (leaf.bold) {
-    children = <strong>{children}</strong>;
-  }
-
-  if (leaf.code) {
-    children = <code>{children}</code>;
-  }
-
-  if (leaf.italic) {
-    children = <em>{children}</em>;
-  }
-
-  if (leaf.underline) {
-    children = <u>{children}</u>;
-  }
-
-  return <span {...attributes}>{children}</span>;
-};
-
-const BlockButton = ({ format, icon }) => {
-  const editor = useSlate();
-  return (
-    <Icon
-      active={isBlockActive(editor, format)}
-      onMouseDown={(event) => {
-        event.preventDefault();
-        toggleBlock(editor, format);
-      }}
-    >
-      <Icon>{icon}</Icon>
-    </Icon>
-  );
-};
-
-const MarkButton = ({ format, icon }) => {
-  const editor = useSlate();
-  return (
-    <Icon
-      active={isMarkActive(editor, format)}
-      onMouseDown={(event) => {
-        event.preventDefault();
-        toggleMark(editor, format);
-      }}
-    >
-      <Icon>{icon}</Icon>
-    </Icon>
-  );
-};
