+++ src/components/Editor/components/toggleButtons.js
@@ -1,0 +1,38 @@
+import { Icon } from "./components";
+import { useSlate } from "slate-react";
+import {
+  isMarkActive,
+  toggleMark,
+  toggleBlock,
+  isBlockActive,
+} from "../editor-utils/toggles";
+
+export const BlockButton = ({ format, icon }) => {
+  const editor = useSlate();
+  return (
+    <Icon
+      active={isBlockActive(editor, format)}
+      onMouseDown={(event) => {
+        event.preventDefault();
+        toggleBlock(editor, format);
+      }}
+    >
+      <Icon>{icon}</Icon>
+    </Icon>
+  );
+};
+
+export const MarkButton = ({ format, icon }) => {
+  const editor = useSlate();
+  return (
+    <Icon
+      active={isMarkActive(editor, format)}
+      onMouseDown={(event) => {
+        event.preventDefault();
+        toggleMark(editor, format);
+      }}
+    >
+      <Icon>{icon}</Icon>
+    </Icon>
+  );
+};
